+++ src/utils/geometry.py
@@ -89,7 +89,7 @@ class Geometry:
 
 class BasicModel(nn.Module):
 
-    def __init__(self, geometry: Geometry, kernel: torch.Tensor = None, trainable_kernel=True, **kwargs):
+    def __init__(self, geometry: Geometry, kernel: torch.Tensor = None, trainable_kernel=True, dtype=torch.complex64, **kwargs):
         "Linear layer consisting of a 1D sinogram kernel in frequency domain"
         super(BasicModel, self).__init__(**kwargs)
         
@@ -100,7 +100,7 @@ class BasicModel(nn.Module):
             #start_kernel = np.linspace(0, 1.0, geometry.fourier_domain.shape[0]) * np.random.triangular(0, 25, 50)
             #if random.random() < 0.5: start_kernel *= -1
             #self.kernel = nn.Parameter(torch.from_numpy(start_kernel).to(DEVICE), requires_grad=trainable_kernel)
-            self.kernel = nn.Parameter(torch.randn(geometry.fourier_domain.shape).to(DEVICE), requires_grad=trainable_kernel)
+            self.kernel = nn.Parameter(torch.randn(geometry.fourier_domain.shape, dtype=dtype).to(DEVICE), requires_grad=trainable_kernel)
         else:
             assert kernel.shape == geometry.fourier_domain.shape, f"wrong formatted specific kernel {kernel.shape} for geometry {geometry}"
             self.kernel = nn.Parameter(kernel.to(DEVICE), requires_grad=trainable_kernel)
